// vizsgamunka/enums/enums.ts
export enum RegistrationStatus {
    Accepted = "Elfogadva",
    Withdrawn = "Visszavonva"
}

export enum ExperienceLevel {
    Beginner = "Kezdő",
    Intermediate = "Haladó",
    Experienced = "Gyakorlott"
}

export enum GameSystem {
    CohorsCthulhu = "Cohors Cthulhu",
    CallOfCthulhu = "Call of Cthulhu",
    D2O = "D2O alapú"
}